Conflicts:
	app/Model/Usuario.php

+++ app/Controller/CatcargosController.php
@@ -66,23 +66,25 @@ class CatcargosController extends AppController {
  * @param string $id
  * @return void
  */
-	public function edit($id = null) {
+	public function editar ($id = null) {
 		if (!$this->Catcargo->exists($id)) {
-			throw new NotFoundException(__('Invalid catcargo'));
-		}
+			$this->Session->setFlash(__('Código de Cargo Invalido.'), array('class' => 'ERROR'));
+		} else {
 		if ($this->request->is(array('post', 'put'))) {
 			if ($this->Catcargo->save($this->request->data)) {
-				$this->Session->setFlash(__('The catcargo has been saved.'));
+				$this->Session->setFlash(__('Se han guardado los cambios con exito.'), array('class' => 'OK'));
 				return $this->redirect(array('action' => 'index'));
 			} else {
-				$this->Session->setFlash(__('The catcargo could not be saved. Please, try again.'));
+				$this->Session->setFlash(__('¡Ha ocurrido un error al guardar los datos! , por favor intente de nuevo.'), array('class' => 'ERROR'));
 			}
 		} else {
 			$options = array('conditions' => array('Catcargo.' . $this->Catcargo->primaryKey => $id));
 			$this->request->data = $this->Catcargo->find('first', $options);
 		}
 	}
-
+         $this->set('title_for_layout', 'Editar');
+        
+        }
 /**
  * delete method
  *
@@ -90,17 +92,26 @@ class CatcargosController extends AppController {
  * @param string $id
  * @return void
  */
-	public function delete($id = null) {
+	public function eliminar($id = null) {
 		$this->Catcargo->id = $id;
 		if (!$this->Catcargo->exists()) {
-			throw new NotFoundException(__('Invalid catcargo'));
+			throw new NotFoundException(__('Código de Cargo Invalido.'), array('class' => 'ERROR'));
 		}
-		$this->request->allowMethod('post', 'delete');
-		if ($this->Catcargo->delete()) {
-			$this->Session->setFlash(__('The catcargo has been deleted.'));
+		if ($this->request->is(array('post', 'delete'))) {
+            if ($this->Catcargo->delete()) {
+			$this->Session->setFlash(__('Se ha eliminado el Cargo con exito'), array('class' => 'OK'));
 		} else {
-			$this->Session->setFlash(__('The catcargo could not be deleted. Please, try again.'));
+			$this->Session->setFlash(__('¡Ha ocurrido un error al eliminar el Cargo! , por favor intente de nuevo.'), array('class' => 'ERROR'));
 		}
 		return $this->redirect(array('action' => 'index'));
 	}
+
+        $options = array('conditions' => array('Cargo.' . $this->Facultade->primaryKey => $id), 'fields' => array('nombre'));
+        $this->set('cargo', $this->Cargo->find('first', $options));
+        $this->set('title_for_layout', 'Eliminar');
+        
+        
+        
+                }
+                
 }
